requirements.txt
# Core dependencies
selenium>=4.0.0
webdriver-manager>=3.8.0
Pillow>=9.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.0
pydantic>=1.8.0

# Optional dependencies
python-multipart  # For file uploads
python-jose[cryptography]  # For JWT support
